use-package helpful
Helpful's presenation of information about key maps is much better.  The vanilla
help buffers are not too inspiring and sometimes contain literal values that are
difficult for the user to interpret.  Helpful is a step in the right direction
towards being able to explore both behavior and the raw state of code.