Introducing loading shim generation
Instead of asking the user to update a list of lisp files to load and including
/ omitting to adjust which modules are in use, allow home manager modules to
insert their lisp file name into a list that posimacs will generate.

The result is a stable init.el and early-init.el even as the user changes which
home manager modules are enabled.  This can compose with independent modules
that use the posimacs.modules option to add their file to the list.