Add Theme
Vterm and several other programs were nicely covered in the doom themes package,
so this addition packages a modified dark theme, Acario dark, shifted towards
official Positron hues, which are pretty vibrant and multicolored.

Currently there is no support for selecting a theme.  This will not arrive until
the dilemmas between home manager, Emacs, and nix are all reconcilled.

This theme is likely only appropriate for trichromes and higher.  Get a hold of
Prot's themes or other built-in Doom themes if you need accessibility standards
compliant themes.  Positron's base color schemes are not going to be easy to
thread into accessibility standards.