Convert to flakes sources (1st pass)
Flakes wants to know about all inputs.

When making a flake compat default.nix, fetchTarball cannot be used, and the
flake-compat input needs to be declared in the flake for use in the default.nix.

Niv style sources also seem to step outside the flake model.

For now, all inputs have been declared upstream in home.nix and are then brought
into each home manager module via specialArgs.

When building flake attributes with `nix build` etc, the system is inferred, but
within a home manager module, we need to provide this information downstream for
looking up the attribute manually.  One solution is to declare the system as a
constant in the module, but this module is definitely going to be used on
multiple systems, and so we must tell the modules from the top level what system
they are on when building the home manager profile.