Updated pmx-ielm-this-buffer
Because this is a really nice workflow for developing lots of elisp in spite of
what the haters say.  You really can't hack on another buffer using just
`eval-last-sexp' in your elisp buffer and `eval-expression' in the target buffer
as quickly.  It's slower to introspect that way.  You can't maintain context for
hacking either.